Fix flac plugin for small files
The flac library does not output decoded data until it has processed a
certain amount of input data or until the end of the stream is reached.
The old version of the code did not take this into account. In
particular, us having inputted all of our data into the flac plugin does
not mean that the flac plugin has outputted all of its data yet.

Hence we remove the `priv->pos == priv-len` check and instead ask the
flac library directly if it's done.